Fix/Add: Make loop class independent
Previously, loop classes were dependent on multiple objects, such
as a logger, tensorboard, and parent loops. This makes it hard to
use loop classes since many requirements are demanded to use them.
Now, one loop class will do one job.

1. `MetricLoop` contains calculated metrics and gives averages of
   those metrics.

2. `ImageLoop` creates a grid image with image tensors.

3. `ProgressbarLoop` will print a progress bar.

Metrics and images are required to be printed in any way. To do
this, each callback in a loop will get a function to process them.
By this design, `MetricLoop` and `ImageLoop` are independent to
any other object.

In addition, python file `writer.py` is changed to `result.py`,
since loop classes in the file are not related to writing something
no more, and are related to handling results of each loop
iteration.